extract model and start using modules
