+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/wdonet/golang-bootcamp-2020
+
+go 1.15
+
+require github.com/gorilla/mux v1.8.0
